+++ python/MaterialX/main.py
@@ -180,10 +180,16 @@ def _getReferencedShaderDefs(self):
     warnings.warn("This function is deprecated; call Material.getShaderNodeDefs instead.", DeprecationWarning, stacklevel = 2)
     return self.getShaderNodeDefs()
 
+def _getReferencingMaterialAssigns(self):
+    "(Deprecated) Return a list of all material assigns that reference this material."
+    warnings.warn("This function is deprecated; call Material.getGeometryBindings instead.", DeprecationWarning, stacklevel = 2)
+    return self.getGeometryBindings()
+
 Material.addOverride = _addOverride
 Material.setOverrideValue = _setOverrideValue
 Material.addShaderRef = _addShaderRef
 Material.getReferencedShaderDefs = _getReferencedShaderDefs
+Material.getReferencingMaterialAssigns = _getReferencingMaterialAssigns
 
 
 #
@@ -240,14 +246,20 @@ GeomInfo.setGeomAttrValue = _setGeomAttrValue
 # Document
 #
 
-def _applyStringSubstitutions(self, filename, geom = '*'):
+def _applyStringSubstitutions(self, filename, geom = '/'):
     """(Deprecated) Given an input filename and geom string, apply any string
         substitutions that have been defined for the given geom to the filename,
         returning the modified filename."""
     warnings.warn("This function is deprecated; call Element.createStringResolver() instead.", DeprecationWarning, stacklevel = 2)
     return self.createStringResolver(geom).resolve(filename, 'filename')
 
+def _generateRequireString(self):
+    """(Deprecated) Generate the require string for a document."""
+    warnings.warn("Require strings are no longer supported in MaterialX.", DeprecationWarning, stacklevel = 2)
+    pass
+
 Document.applyStringSubstitutions = _applyStringSubstitutions
+Document.generateRequireString = _generateRequireString
 
 
 #