+++ source/MaterialXCore/Element.cpp
@@ -18,8 +18,9 @@ const string Element::FILE_PREFIX_ATTRIBUTE = "fileprefix";
 const string Element::GEOM_PREFIX_ATTRIBUTE = "geomprefix";
 const string Element::COLOR_SPACE_ATTRIBUTE = "colorspace";
 const string Element::TARGET_ATTRIBUTE = "target";
+const string Element::INHERIT_ATTRIBUTE = "inherit";
+const string Element::NAMESPACE_ATTRIBUTE = "namespace";
 const string ValueElement::VALUE_ATTRIBUTE = "value";
-const string ValueElement::PUBLIC_NAME_ATTRIBUTE = "publicname";
 const string ValueElement::INTERFACE_NAME_ATTRIBUTE = "interfacename";
 const string ValueElement::IMPLEMENTATION_NAME_ATTRIBUTE = "implname";
 
@@ -288,6 +289,18 @@ ConstElementPtr Element::getRoot() const
     return root;
 }
 
+bool Element::hasInheritedBase(ConstElementPtr base) const
+{
+    for (ConstElementPtr elem : traverseInheritance())
+    {
+        if (elem == base)
+        {
+            return true;
+        }
+    }
+    return false;
+}
+
 bool Element::hasInheritanceCycle() const
 {
     try
@@ -376,6 +389,11 @@ bool Element::validate(string* message) const
     {
         validateRequire(getDocument()->hasColorManagementSystem(), res, message, "Colorspace set without color management system");
     }
+    if (hasInheritString())
+    {
+        bool validInherit = getInheritsFrom() && getInheritsFrom()->getCategory() == getCategory();
+        validateRequire(validInherit, res, message, "Invalid element inheritance");
+    }
     for (ElementPtr child : getChildren())
     {
         res = child->validate(message) && res;
@@ -399,12 +417,12 @@ StringResolverPtr Element::createStringResolver(const string& geom) const
         ConstDocumentPtr doc = getDocument();
         for (GeomInfoPtr geomInfo : doc->getGeomInfos())
         {
-            if (!geomStringsMatch(geom, geomInfo->getGeom()))
+            if (!geomStringsMatch(geom, geomInfo->getActiveGeom()))
                 continue;
-            for (GeomAttrPtr geomAttr : geomInfo->getGeomAttrs())
+            for (TokenPtr token : geomInfo->getTokens())
             {
-                string key = "%" + geomAttr->getName();
-                string value = geomAttr->getResolvedValueString();
+                string key = "%" + token->getName();
+                string value = token->getResolvedValueString();
                 resolver->setFilenameSubstitution(key, value);
             }
         }
@@ -480,7 +498,7 @@ ValuePtr ValueElement::getDefaultValue() const
         NodeDefPtr decl = getParent()->asA<InterfaceElement>()->getDeclaration();
         if (decl)
         {
-            ValueElementPtr value = decl->getChildOfType<ValueElement>(getName());
+            ValueElementPtr value = decl->getActiveValueElement(getName());
             if (value)
             {
                 return value->getValue();
@@ -584,8 +602,6 @@ INSTANTIATE_SUBCLASS(T)
 INSTANTIATE_CONCRETE_SUBCLASS(BindParam, "bindparam")
 INSTANTIATE_CONCRETE_SUBCLASS(BindInput, "bindinput")
 INSTANTIATE_CONCRETE_SUBCLASS(Collection, "collection")
-INSTANTIATE_CONCRETE_SUBCLASS(CollectionAdd, "collectionadd")
-INSTANTIATE_CONCRETE_SUBCLASS(CollectionRemove, "collectionremove")
 INSTANTIATE_CONCRETE_SUBCLASS(Document, "materialx")
 INSTANTIATE_CONCRETE_SUBCLASS(GenericElement, "generic")
 INSTANTIATE_CONCRETE_SUBCLASS(GeomAttr, "geomattr")
@@ -593,15 +609,12 @@ INSTANTIATE_CONCRETE_SUBCLASS(GeomInfo, "geominfo")
 INSTANTIATE_CONCRETE_SUBCLASS(Implementation, "implementation")
 INSTANTIATE_CONCRETE_SUBCLASS(Input, "input")
 INSTANTIATE_CONCRETE_SUBCLASS(Look, "look")
-INSTANTIATE_CONCRETE_SUBCLASS(LookInherit, "lookinherit")
 INSTANTIATE_CONCRETE_SUBCLASS(Material, "material")
 INSTANTIATE_CONCRETE_SUBCLASS(MaterialAssign, "materialassign")
-INSTANTIATE_CONCRETE_SUBCLASS(MaterialInherit, "materialinherit")
 INSTANTIATE_CONCRETE_SUBCLASS(Member, "member")
 INSTANTIATE_CONCRETE_SUBCLASS(Node, "node")
 INSTANTIATE_CONCRETE_SUBCLASS(NodeDef, "nodedef")
 INSTANTIATE_CONCRETE_SUBCLASS(NodeGraph, "nodegraph")
-INSTANTIATE_CONCRETE_SUBCLASS(Override, "override")
 INSTANTIATE_CONCRETE_SUBCLASS(Output, "output")
 INSTANTIATE_CONCRETE_SUBCLASS(Parameter, "parameter")
 INSTANTIATE_CONCRETE_SUBCLASS(Property, "property")
@@ -609,7 +622,11 @@ INSTANTIATE_CONCRETE_SUBCLASS(PropertyAssign, "propertyassign")
 INSTANTIATE_CONCRETE_SUBCLASS(PropertySet, "propertyset")
 INSTANTIATE_CONCRETE_SUBCLASS(PropertySetAssign, "propertysetassign")
 INSTANTIATE_CONCRETE_SUBCLASS(ShaderRef, "shaderref")
+INSTANTIATE_CONCRETE_SUBCLASS(Token, "token")
 INSTANTIATE_CONCRETE_SUBCLASS(TypeDef, "typedef")
+INSTANTIATE_CONCRETE_SUBCLASS(Variant, "variant")
+INSTANTIATE_CONCRETE_SUBCLASS(VariantAssign, "variantassign")
+INSTANTIATE_CONCRETE_SUBCLASS(VariantSet, "variantset")
 INSTANTIATE_CONCRETE_SUBCLASS(Visibility, "visibility")
 
 } // namespace MaterialX