+++ source/MaterialXCore/Material.h
@@ -11,6 +11,7 @@
 
 #include <MaterialXCore/Library.h>
 
+#include <MaterialXCore/Geom.h>
 #include <MaterialXCore/Node.h>
 #include <MaterialXCore/Value.h>
 
@@ -24,14 +25,18 @@ using ConstMaterialPtr = shared_ptr<const class Material>;
 
 /// A shared pointer to a ShaderRef
 using ShaderRefPtr = shared_ptr<class ShaderRef>;
+/// A shared pointer to a const ShaderRef
+using ConstShaderRefPtr = shared_ptr<const class ShaderRef>;
+
 /// A shared pointer to a BindParam
 using BindParamPtr = shared_ptr<class BindParam>;
+/// A shared pointer to a const BindParam
+using ConstBindParamPtr = shared_ptr<const class BindParam>;
+
 /// A shared pointer to a BindInput
 using BindInputPtr = shared_ptr<class BindInput>;
-/// A shared pointer to an Override
-using OverridePtr = shared_ptr<class Override>;
-/// A shared pointer to a MaterialInherit
-using MaterialInheritPtr = shared_ptr<class MaterialInherit>;
+/// A shared pointer to a const BindInput
+using ConstBindInputPtr = shared_ptr<const class BindInput>;
 
 /// @class Material
 /// A material element within a Document.
@@ -88,91 +93,6 @@ class Material : public Element
     }
 
     /// @}
-    /// @name Override Elements
-    /// @{
-
-    /// Add a Override to the material.
-    /// @param name The name of the new Override.
-    ///     If no name is specified, then a unique name will automatically be
-    ///     generated.
-    /// @return A shared pointer to the new Override.
-    OverridePtr addOverride(const string& name)
-    {
-        return addChild<Override>(name);
-    }
-
-    /// Return the Override, if any, with the given name.
-    OverridePtr getOverride(const string& name) const
-    {
-        return getChildOfType<Override>(name);
-    }
-
-    /// Return a vector of all Override elements that belong to this material,
-    /// taking material inheritance into account.
-    vector<OverridePtr> getActiveOverrides() const;
-
-    /// Return a vector of all Override elements 
-    vector<OverridePtr> getOverrides() const
-    {
-        return getChildrenOfType<Override>();
-    }
-
-    /// Remove the Override, if any, with the given name.
-    void removeOverride(const string& name)
-    {
-        removeChildOfType<Override>(name);
-    }
-
-    /// Set the value of an override by its name, creating a child element
-    /// to hold the override if needed.
-    template<class T> OverridePtr setOverrideValue(const string& name,
-                                                   const T& value,
-                                                   const string& type = EMPTY_STRING);
-
-    /// @}
-    /// @name MaterialInherit Elements
-    /// @{
-
-    /// Add a MaterialInherit to the material.
-    /// @param name The name of the new MaterialInherit.
-    ///     If no name is specified, then a unique name will automatically be
-    ///     generated.
-    /// @return A shared pointer to the new MaterialInherit.
-    MaterialInheritPtr addMaterialInherit(const string& name = EMPTY_STRING)
-    {
-        MaterialInheritPtr ptr = addChild<MaterialInherit>(name);
-        return ptr;
-    }
-
-    /// Return the MaterialInherit, if any, with the given name.
-    MaterialInheritPtr getMaterialInherit(const string& name) const
-    {
-        return getChildOfType<MaterialInherit>(name);
-    }
-
-    /// Return a vector of all MaterialInherit elements in the material.
-    vector<MaterialInheritPtr> getMaterialInherits() const
-    {
-        return getChildrenOfType<MaterialInherit>();
-    }
-
-    /// Remove the MaterialInherit, if any, with the given name.
-    void removeMaterialInherit(const string& name)
-    {
-        removeChildOfType<MaterialInherit>(name);
-    }
-
-    /// @}
-    /// @name Inheritance
-    /// @{
-
-    /// Set the material element that this one inherits from.
-    void setInheritsFrom(ElementPtr mat) override;
-
-    /// Return the material element, if any, that this one inherits from.
-    ElementPtr getInheritsFrom() const override;
-
-    /// @}
     /// @name NodeDef References
     /// @{
 
@@ -185,13 +105,6 @@ class Material : public Element
     /// @return A vector of shared pointers to NodeDef elements.
     vector<NodeDefPtr> getShaderNodeDefs(const string& target = EMPTY_STRING,
                                          const string& type = EMPTY_STRING) const;
-
-    /// @}
-    /// @name MaterialAssign References
-    /// @{
-
-    /// Return all MaterialAssign elements that reference this material.
-    vector<MaterialAssignPtr> getReferencingMaterialAssigns() const;
 
     /// @}
     /// @name Primary Shader
@@ -253,17 +166,12 @@ class Material : public Element
     /// @name Geometry Bindings
     /// @{
 
-    /// Return all geometry strings that are bound to this material by Look
-    /// elements.  Note that this method only considers geometry strings,
-    /// not geometric collections.
-    /// @return A vector of geometry strings, each containing an array of
-    ///    geom names.
-    vector<string> getBoundGeomStrings() const;
-
-    /// Return all geometry collections that are bound to this material by
-    /// Look elements.
-    /// @return A vector of shared pointers to Collection elements.
-    vector<CollectionPtr> getBoundGeomCollections() const;
+    /// Return a vector of all MaterialAssign elements that bind this material
+    /// to the given geometry string.
+    /// @param geom The geometry for which material bindings should be returned.
+    ///    By default, this argument is the universal geometry string "/", and
+    ///    all material bindings are returned.
+    vector<MaterialAssignPtr> getGeometryBindings(const string& geom = UNIVERSAL_GEOM_NAME) const;
 
     /// @}
     /// @name Validation
@@ -348,7 +256,7 @@ class BindInput : public ValueElement
     /// @{
 
     /// Set the output to which the BindInput is connected.
-    void setConnectedOutput(OutputPtr output);
+    void setConnectedOutput(ConstOutputPtr output);
 
     /// Return the output, if any, to which the BindInput is connected.
     OutputPtr getConnectedOutput() const;
@@ -546,58 +454,6 @@ class ShaderRef : public Element
     static const string NODE_ATTRIBUTE;
     static const string NODE_DEF_ATTRIBUTE;
 };
-
-/// @class Override
-/// An override element within a Material.
-///
-/// An Override modifies the uniform value of a public Parameter or Input
-/// within the scope of a Material.
-class Override : public ValueElement
-{
-  public:
-    Override(ElementPtr parent, const string& name) :
-        ValueElement(parent, CATEGORY, name)
-    {
-    }
-    virtual ~Override() { }
-
-    /// @name Connections
-    /// @{
-
-    /// Return the element, if any, that is modified by this override.
-    ConstElementPtr getReceiver() const;
-
-    /// @}
-
-  public:
-    static const string CATEGORY;
-};
-
-/// @class MaterialInherit
-/// A material inheritance element within a Material.
-class MaterialInherit : public Element
-{
-  public:
-    MaterialInherit(ElementPtr parent, const string& name) :
-        Element(parent, CATEGORY, name)
-    {
-    }
-    virtual ~MaterialInherit() { }
-
-  public:
-    static const string CATEGORY;
-};
-
-template<class T> OverridePtr Material::setOverrideValue(const string& name,
-                                                         const T& value,
-                                                         const string& type)
-{
-    OverridePtr override = getChildOfType<Override>(name);
-    if (!override)
-        override = addChild<Override>(name);
-    override->setValue(value, type);
-    return override;
-}
 
 } // namespace MaterialX