+++ source/MaterialXTest/Geom.cpp
@@ -9,23 +9,38 @@
 
 namespace mx = MaterialX;
 
-TEST_CASE("Geom", "[geom]")
+TEST_CASE("Geom strings", "[geom]")
+{
+    // Test for overlapping paths.
+    REQUIRE(mx::geomStringsMatch("/", "/robot1"));
+    REQUIRE(mx::geomStringsMatch("/robot1", "/robot1/left_arm"));
+    REQUIRE(mx::geomStringsMatch("/robot1, /robot2", "/robot2/left_arm"));
+    REQUIRE(!mx::geomStringsMatch("", "/robot1"));
+    REQUIRE(!mx::geomStringsMatch("/robot1", "/robot2"));
+    REQUIRE(!mx::geomStringsMatch("/robot1, /robot2", "/robot3"));
+
+    // Test that one path contains another.
+    REQUIRE(mx::geomStringsMatch("/", "/robot1", true));
+    REQUIRE(!mx::geomStringsMatch("/robot1", "/", true));
+}
+
+TEST_CASE("Geom elements", "[geom]")
 {
     mx::DocumentPtr doc = mx::createDocument();
 
-    // Add geominfos and geomattrs
-    mx::GeomInfoPtr geominfo1 = doc->addGeomInfo("geominfo1", "/robot1,/robot2");
-    geominfo1->setGeomAttrValue("asset", std::string("robot"));
+    // Add geominfos and tokens
+    mx::GeomInfoPtr geominfo1 = doc->addGeomInfo("geominfo1", "/robot1, /robot2");
+    geominfo1->setTokenValue("asset", std::string("robot"));
     mx::GeomInfoPtr geominfo2 = doc->addGeomInfo("geominfo2", "/robot1");
-    geominfo2->setGeomAttrValue("id", std::string("01"));
+    geominfo2->setTokenValue("id", std::string("01"));
     mx::GeomInfoPtr geominfo3 = doc->addGeomInfo("geominfo3", "/robot2");
-    geominfo3->setGeomAttrValue("id", std::string("02"));
-    REQUIRE_THROWS_AS(doc->addGeomInfo("geominfo1"), mx::Exception);
+    geominfo3->setTokenValue("id", std::string("02"));
+    REQUIRE_THROWS_AS(doc->addGeomInfo("geominfo1"), mx::Exception&);
 
     // Create a node graph with a single image node.
     mx::NodeGraphPtr nodeGraph = doc->addNodeGraph();
     nodeGraph->setFilePrefix("folder/");
-    REQUIRE_THROWS_AS(doc->addNodeGraph(nodeGraph->getName()), mx::Exception);
+    REQUIRE_THROWS_AS(doc->addNodeGraph(nodeGraph->getName()), mx::Exception&);
     mx::NodePtr image = nodeGraph->addNode("image");
     image->setParameterValue("file", std::string("%asset%id_diffuse_%UDIM.tif"), mx::FILENAME_TYPE_STRING);
 
@@ -38,13 +53,27 @@ TEST_CASE("Geom", "[geom]")
     REQUIRE(fileParam->getResolvedValueString(resolver1) == "folder/robot01_diffuse_1001.tif");
     REQUIRE(fileParam->getResolvedValueString(resolver2) == "folder/robot02_diffuse_1002.tif");
 
+    // Create a base collection.
+    mx::CollectionPtr collection1 = doc->addCollection("collection1");
+    collection1->setIncludeGeom("/scene1");
+    collection1->setExcludeGeom("/scene1/sphere2");
+    REQUIRE(collection1->matchesGeomString("/scene1/sphere1"));
+    REQUIRE(!collection1->matchesGeomString("/scene1/sphere2"));
+
+    // Create a derived collection.
+    mx::CollectionPtr collection2 = doc->addCollection("collection2");
+    collection2->setIncludeCollection(collection1);
+    REQUIRE(collection2->matchesGeomString("/scene1/sphere1"));
+    REQUIRE(!collection2->matchesGeomString("/scene1/sphere2"));
+
+    // Create and test an include cycle.
+    collection1->setIncludeCollection(collection2);
+    REQUIRE(!doc->validate());
+    collection1->setIncludeCollection(nullptr);
+    REQUIRE(doc->validate());
+
     // Test geometry string substitutions.
-    mx::CollectionPtr collection = doc->addCollection("collection1");
-    mx::CollectionAddPtr collectionAdd = collection->addCollectionAdd("collectionAdd1");
-    std::string geom = "|group1|sphere1";
-    collectionAdd->setGeom(geom);
-    collectionAdd->setGeomPrefix("/geomPrefix1");
-    mx::StringResolverPtr resolver3 = collectionAdd->createStringResolver();
-    resolver3->setGeomNameSubstitution("|", "/");
-    REQUIRE(resolver3->resolve(collectionAdd->getGeom(), mx::GEOMNAME_TYPE_STRING) == "/geomPrefix1/group1/sphere1");
+    collection1->setGeomPrefix("/root");
+    REQUIRE(collection1->matchesGeomString("/root/scene1"));
+    REQUIRE(!collection1->matchesGeomString("/root/scene2"));
 }